add+fill Material file, helpers.py, rename SWT.py
